Getting basic functionality working outside of the main controller.
