+++ stash_engine/lib/stash_engine.rb
@@ -28,6 +28,7 @@ module StashEngine
     config.autoload_paths << File.expand_path('lib/stash/doi', __dir__)
     config.autoload_paths << File.expand_path('lib/stash/indexer', __dir__)
     config.autoload_paths << File.expand_path('lib/stash/payments', __dir__)
+    config.autoload_paths << File.expand_path('lib/stash/download', __dir__)
 
     # :nocov:
     initializer :append_migrations do |app|
